requirements.txt
# Core dependencies
click>=8.0.0
requests>=2.25.0
pillow>=9.0.0
pyyaml>=6.0
openai>=1.0.0
jsonschema>=4.0.0
python-dotenv>=0.19.0
opencv-python>=4.5.0
numpy>=1.20.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0

# Development dependencies
black>=23.0.0
isort>=5.10.0
flake8>=6.0.0
mypy>=1.0.0